dto에 있는 필드 중 int64 타입을 int32로 변경합니다.

+++ src/main/java/com/hwans/apiserver/dto/blog/PostDto.java
@@ -30,5 +30,5 @@ public class PostDto implements Serializable {
     @ApiModelProperty(value = "댓글")
     Set<SimpleCommentDto> comments;
     @ApiModelProperty(value = "좋아요 수")
-    Long likeCount;
+    int likeCount;
 }
